project/requirements.txt
celery==5.2.3
Django==4.0
djangorestframework==3.12.4
redis==4.1.4
requests==2.25.1
psycopg2==2.8.6
drf-yasg==1.20.0
